backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pandas>=2.0.0
numpy>=1.24.0
openai>=1.0.0
plotly>=5.18.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
google-generativeai>=0.3.0
scipy>=1.11.0
